# Minimize allocations when mangling forwarded header.