--- a/pkg/queue/forwarded_shim_test.go
+++ b/pkg/queue/forwarded_shim_test.go
@@ -69,6 +69,10 @@ func TestForwardedShimHandler(t *testing.T) {
 		xfp:  "p",
 		want: "host=h;proto=p",
 	}, {
+		name: "broken for",
+		xff:  ",,,",
+		want: "for=, for=, for=",
+	}, {
 		name: "existing fwd",
 		xff:  "127.0.0.1, ::1",
 		xfh:  "h",