<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

**Note:** This is totally premature optimization. This is likely not an issue in production at all.

I tried this to have a little fun with writing code that uses as little allocations as possible. Just for fun mostly. Ended up with something I kinda liked and the benchmark difference is awesome. Sending this in for discussion thus anyway. Less allocations is always better, right?

### Benchmark results

```
benchmark                            old ns/op     new ns/op     delta
BenchmarkForwardedShimHandler-12     846           134           -84.16%
benchmark                            old allocs     new allocs     delta
BenchmarkForwardedShimHandler-12     14             1              -92.86%
benchmark                            old bytes     new bytes     delta
BenchmarkForwardedShimHandler-12     480           96            -80.00%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov @tanzeeb