@tanzeeb 2020-04-23
This is great! This code is called on every request so each nanosecond counts!

/lgtm